Refactor imports to use parseTypes.js for JsonSchema7Type
