+++ src/Options.ts
@@ -1,6 +1,6 @@
 import { ZodSchema, ZodTypeDef } from "zod";
 import { Refs, Seen } from "./Refs";
-import { JsonSchema7Type } from "./parseDef";
+import { JsonSchema7Type } from "./parseTypes";
 
 export type Targets = "jsonSchema7" | "jsonSchema2019-09" | "openApi3" | "openAi";
