+++ src/parsers/array.ts
@@ -1,6 +1,7 @@
 import { ZodArrayDef, ZodFirstPartyTypeKind } from "zod";
 import { ErrorMessages, setResponseValueAndErrors } from "../errorMessages.js";
-import { JsonSchema7Type, parseDef } from "../parseDef.js";
+import { parseDef } from "../parseDef.js";
+import { JsonSchema7Type } from "../parseTypes.js";
 import { Refs } from "../Refs.js";
 
 export type JsonSchema7ArrayType = {
