+++ src/parsers/effects.ts
@@ -1,5 +1,6 @@
 import { ZodEffectsDef } from "zod";
-import { JsonSchema7Type, parseDef } from "../parseDef.js";
+import { parseDef } from "../parseDef.js";
+import { JsonSchema7Type } from "../parseTypes.js";
 import { Refs } from "../Refs.js";
 
 export function parseEffectsDef(
