+++ src/parsers/optional.ts
@@ -1,5 +1,6 @@
 import { ZodOptionalDef } from "zod";
-import { JsonSchema7Type, parseDef } from "../parseDef.js";
+import { parseDef } from "../parseDef.js";
+import { JsonSchema7Type } from "../parseTypes.js";
 import { Refs } from "../Refs.js";
 
 export const parseOptionalDef = (
