+++ src/parsers/pipeline.ts
@@ -1,5 +1,6 @@
 import { ZodPipelineDef } from "zod";
-import { JsonSchema7Type, parseDef } from "../parseDef.js";
+import { parseDef } from "../parseDef.js";
+import { JsonSchema7Type } from "../parseTypes.js";
 import { Refs } from "../Refs.js";
 import { JsonSchema7AllOfType } from "./intersection.js";
