+++ src/parsers/record.ts
@@ -4,7 +4,8 @@ import {
   ZodRecordDef,
   ZodTypeAny,
 } from "zod";
-import { JsonSchema7Type, parseDef } from "../parseDef.js";
+import { parseDef } from "../parseDef.js";
+import { JsonSchema7Type } from "../parseTypes.js";
 import { Refs } from "../Refs.js";
 import { JsonSchema7EnumType } from "./enum.js";
 import { JsonSchema7ObjectType } from "./object.js";
