+++ src/parsers/tuple.ts
@@ -1,5 +1,6 @@
 import { ZodTupleDef, ZodTupleItems, ZodTypeAny } from "zod";
-import { JsonSchema7Type, parseDef } from "../parseDef.js";
+import { parseDef } from "../parseDef.js";
+import { JsonSchema7Type } from "../parseTypes.js";
 import { Refs } from "../Refs.js";
 
 export type JsonSchema7TupleType = {
