+++ src/parsers/union.ts
@@ -4,7 +4,8 @@ import {
   ZodTypeAny,
   ZodUnionDef,
 } from "zod";
-import { JsonSchema7Type, parseDef } from "../parseDef.js";
+import { parseDef } from "../parseDef.js";
+import { JsonSchema7Type } from "../parseTypes.js";
 import { Refs } from "../Refs.js";
 
 export const primitiveMappings = {
